+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Added 
 - added improves wifideauth module for support multi targets 
 - added improves ConsoleUi and added help for command add, rm from wifideauth module 
+- added hostapd configuration file from wifipumpkin3 console
 
 ### Changed